studio 0

+++ src/studio0/BurritoBudget.java
@@ -3,12 +3,12 @@ package studio0;
 public class BurritoBudget {
 
 		public static void main(String args[]) {
-		int burritosEatenWeek1 = 7;
-		int burritosEatenWeek2 = 4;
-		int burritosEatenWeek3 = 19; //don't judge me
-		int burritosEatenWeek4 = 1;
+		double burritosEatenWeek1 = 7;
+		double burritosEatenWeek2 = 4;
+		double burritosEatenWeek3 = 19; //don't judge me
+		double burritosEatenWeek4 = 1;
 		
-		int totalBurritos = burritosEatenWeek1 + burritosEatenWeek2 + burritosEatenWeek3 + burritosEatenWeek4;
+		double totalBurritos = burritosEatenWeek1 + burritosEatenWeek2 + burritosEatenWeek3 + burritosEatenWeek4;
 		double averageWeeklyBurritos = totalBurritos / 4;
 		
 		double burritoPrice = 7.50;
